Refactor observability components: Renamed reporting and tracing to exporters and logs, respectively
