+++ SimNextgenApp.Demo/Scenarios/SimpleGenerator.cs
@@ -6,8 +6,8 @@ using SimNextgenApp.Core.Strategies;
 using SimNextgenApp.Core.Utilities;
 using SimNextgenApp.Demo.CustomModels;
 using SimNextgenApp.Modeling.Generator;
-using SimNextgenApp.Observability.Tracing;
-using SimNextgenApp.Observability.Reporting;
+using SimNextgenApp.Observability.Logs;
+using SimNextgenApp.Observability.Exporters;
 
 namespace SimNextgenApp.Demo.Scenarios;
