+++ SimNextgenApp.Demo/Scenarios/SimpleServerAndGenerator.cs
@@ -4,9 +4,9 @@ using SimNextgenApp.Core;
 using SimNextgenApp.Core.Strategies;
 using SimNextgenApp.Core.Utilities;
 using SimNextgenApp.Demo.CustomModels;
-using SimNextgenApp.Observability.Tracing;
+using SimNextgenApp.Observability.Logs;
 using SimNextgenApp.Observability.Metrics;
-using SimNextgenApp.Observability.Reporting;
+using SimNextgenApp.Observability.Exporters;
 
 namespace SimNextgenApp.Demo.Scenarios;
