+++ SimNextgenApp.Tests/Core/SimulationEngineTests.cs
@@ -6,7 +6,7 @@ using SimNextgenApp.Core.Utilities;
 using SimNextgenApp.Events;
 using SimNextgenApp.Exceptions;
 using SimNextgenApp.Modeling;
-using SimNextgenApp.Observability.Tracing;
+using SimNextgenApp.Observability.Logs;
 
 namespace SimNextgenApp.Tests.Core;
